relocate game view-state types to the game package and decouple Player from HTTP
