+++ game-service-server/src/test/scala/com/andy327/server/actors/battleship/BattleshipActorSpec.scala
@@ -12,7 +12,7 @@ import com.andy327.model.core.{GameError, PlayerId}
 import com.andy327.server.actors.core.{GameManager, PlayerActor, PlayerEvent, TurnBasedGameActor}
 import com.andy327.server.actors.persistence.PersistenceProtocol
 import com.andy327.server.analytics.NoOpAnalyticsPublisher
-import com.andy327.server.http.json.{BattleshipState, GameState}
+import com.andy327.server.game.{BattleshipState, GameState}
 
 class BattleshipActorSpec extends AnyWordSpecLike with Matchers {
   private val testKit = ActorTestKit()
