+++ game-service-server/src/test/scala/com/andy327/server/http/json/BattleshipStateSpec.scala
@@ -7,6 +7,7 @@ import org.scalatest.wordspec.AnyWordSpec
 
 import com.andy327.model.battleship.{Battleship, Coord, Player1, Player2, PlayerBoard, Seat, Ship}
 import com.andy327.model.core.PlayerId
+import com.andy327.server.game.BattleshipState
 
 class BattleshipStateSpec extends AnyWordSpec with Matchers {
   val alice: PlayerId = UUID.randomUUID()
